move DMOD specifics to proxy mixin
