+++ python/gui/MaaS/cbv/utils.py
@@ -1,0 +1,28 @@
+def extract_log_data(kwargs):
+    """
+    Extracts standard names messages from kwargs
+
+    :param kwargs: A dictionary of named arguments
+
+    """
+    # If a list of error messages wasn't passed, create one
+    if 'errors' not in kwargs:
+        errors = list()
+    else:
+        # Otherwise continue to use the passed in list
+        errors = kwargs['errors']  # type: list
+
+    # If a list of warning messages wasn't passed create one
+    if 'warnings' not in kwargs:
+        warnings = list()
+    else:
+        # Otherwise continue to use the passed in list
+        warnings = kwargs['warnings']  # type: list
+
+    # If a list of basic messages wasn't passed, create one
+    if 'info' not in kwargs:
+        info = list()
+    else:
+        # Otherwise continue to us the passed in list
+        info = kwargs['info']  # type: list
+    return errors, warnings, info
